Move ConversionFlag to ast
